+++ JenAI/src/main/java/dev/ceccon/client/LLMSanitizer.java
@@ -81,6 +81,11 @@ public class LLMSanitizer {
 
         sanitized = sanitized.replaceAll("–", "-");
 
+        sanitized = sanitized.replaceAll("’", "'");
+        sanitized = sanitized.replaceAll("‘", "'");
+        sanitized = sanitized.replaceAll("“", "\"");
+        sanitized = sanitized.replaceAll("”", "\"");
+
         return sanitized;
     }
 }